+++ components/ota_upload/ota_upload_mdns.c
@@ -1,15 +1,15 @@
 //-- SPDX-License-Identifier: GPL-3.0-or-later
 //-- Copyright (C) 2026 Willem Aandewiel
 
-#include "ota_manager_internal.h"
+#include "ota_upload_internal.h"
 
 #include "esp_log.h"
 #include "mdns.h"
 
-static const char *tag = "ota_manager_mdns";
+static const char *tag = "ota_upload_mdns";
 static bool mdns_started;
 
-esp_err_t ota_manager_mdns_start(const char *hostname, uint16_t port)
+esp_err_t ota_upload_mdns_start(const char *hostname, uint16_t port)
 {
   if (mdns_started)
   {
@@ -33,7 +33,7 @@ esp_err_t ota_manager_mdns_start(const char *hostname, uint16_t port)
     return err;
   }
 
-  err = mdns_instance_name_set("ESP-IDF OTA Manager");
+  err = mdns_instance_name_set("ESP-IDF OTA Upload");
   if (err != ESP_OK)
   {
     ESP_LOGE(tag, "Failed to set instance name: %s", esp_err_to_name(err));
@@ -76,7 +76,7 @@ esp_err_t ota_manager_mdns_start(const char *hostname, uint16_t port)
   return ESP_OK;
 }
 
-void ota_manager_mdns_stop(void)
+void ota_upload_mdns_stop(void)
 {
   if (!mdns_started)
   {